django/requirements.txt
Django==3.0.4
django-postgres-copy==2.4.2
django-createsuperuser==2019.9.7
django-extensions==2.2.9
django-redis==4.11.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
gunicorn==20.0.4
redis==3.4.1